move to RscTitles, hide indicator on vehicle
