+++ addons/ui/XEH_PREP.hpp
@@ -1,5 +1,7 @@
 PREP(compileConfigUI);
+PREP(createSpeedIndicator);
 PREP(handlePlayerChanged);
+PREP(handlePlayerVehicle);
 PREP(handleSpeedIndicator);
 PREP(moduleInit);
 PREP(onAnimChanged);
